Version bump for packages listed in setup.py
