+++ .github/workflows/build.yml
@@ -1,9 +1,7 @@
 name: SAIL CI
 on: 
   push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+    branches: [ version-bump ]
 
 jobs:
   build:
