+++ requirements.txt
@@ -1,16 +1,3 @@
-setuptools
-matplotlib
-logzero
-ipykernel
-dill==0.3.5.1
-numpy>=1.21.0
-scipy>=1.5.2
-river==0.10.1
-pandas>=1.3.0
-ray==1.11.0
-scikit-learn>=0.24.2
-tensorflow==2.9.1
-torch==1.10.2
-skorch==0.11.0
-scikeras==0.6.1
-scikit-multiflow==0.5.3
+ # installs dependencies from ./setup.py, and the package itself,
+ # in editable mode
+ -e .[all]
