+++ requirements_bkp.txt
@@ -1,0 +1,16 @@
+setuptools
+matplotlib
+logzero
+ipykernel
+dill
+numpy==1.24.*
+scipy==1.10.*
+pandas==1.5.*
+scikit-learn==1.2.*
+river==0.15.*
+ray==2.2.*
+tensorflow==2.10.1
+torch==1.12.0
+skorch==0.12.1
+scikeras==0.10.0
+scikit-multiflow==0.5.3
