+++ sail/__init__.py
@@ -1,5 +1,3 @@
-from importlib_metadata import version
-
 
 __version__ = "0.1.1"
 __package_name__ = "SAIL"
